requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
celery-once==3.0.1
celery-redbeat==2.0.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.6
defusedxml==0.6.0
Django==3.1.6
django-allauth==0.44.0
django-cors-headers==3.7.0
django-jazzmin==2.4.4
django-post-office==3.5.3
django-timezone-field==4.1.1
django-sekizai==2.0.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
idna==2.10
jsonfield==3.1.0
kombu==5.0.2
mysqlclient==2.0.3
oauthlib==3.1.0
prompt-toolkit==3.0.16
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
raven==6.10.0
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
tenacity==6.3.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
